# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RealSense Camera API",
    "version": "1.0.0",
    "description": "A simple RealSense Camera API"
  },
  "paths": {
...
    "/aas/ContactInformation": {
      "get": {
        "summary": "Get contact information",
        "responses": {
          "200": {
            "description": "Contact information retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "email": {
                      "type": "string",
                      "example": "johndoe@example.com"
                    },
                    "phone": {
                      "type": "string",
                      "example": "+1234567890"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact information not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/aas/SoftwareNamePlate": {
      "get": {
        "summary": "Get software nameplate",
        "responses": {
          "200": {
            "description": "Software nameplate retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "software_version": {
                      "type": "string",
                      "example": "1.0.0"
                    },
                    "software_name": {
                      "type": "string",
                      "example": "RealSense Camera API"
                    },
                    "developer": {
                      "type": "string",
                      "example": "YourName"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Software nameplate not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}